chore: 디펜던시 에러 수정
