+++ src/basic/__tests__/basic.test_origin.js
@@ -1,5 +1,5 @@
-import userEvent from '@testing-library/user-event';
-import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import { beforeEach, afterEach, describe, expect, it, vi } from "vitest";
+import userEvent from "@testing-library/user-event";
 
 describe('basic 테스트', () => {
   // 공통 헬퍼 함수
@@ -28,13 +28,13 @@ describe('basic 테스트', () => {
   };
 
   describe.each([
-    { type: 'origin', loadFile: () => import('../../main.original.js') },
-    { type: 'basic', loadFile: () => import('../main.basic.js') },
+    { type: 'origin', loadFile: () => import('../../main.original.js'), },
+    { type: 'basic', loadFile: () => import('../main.basic.js'), },
   ])('$type 장바구니 상세 기능 테스트', ({ loadFile }) => {
     let sel, addBtn, cartDisp, sum, stockInfo, itemCount, loyaltyPoints, discountInfo;
 
     beforeEach(async () => {
-      vi.setSystemTime(new Date('2025-07-28'));
+
       vi.spyOn(window, 'alert').mockImplementation(() => {});
 
       // 전체 DOM 재초기화
@@ -56,7 +56,6 @@ describe('basic 테스트', () => {
     });
 
     afterEach(() => {
-      vi.setSystemTime(new Date('2025-07-28'));
       vi.restoreAllMocks();
     });
 
@@ -69,7 +68,7 @@ describe('basic 테스트', () => {
             { id: 'p2', name: '생산성 폭발 마우스', price: '20000원', stock: 30, discount: 15 },
             { id: 'p3', name: '거북목 탈출 모니터암', price: '30000원', stock: 20, discount: 20 },
             { id: 'p4', name: '에러 방지 노트북 파우치', price: '15000원', stock: 0, discount: 5 },
-            { id: 'p5', name: '코딩할 때 듣는 Lo-Fi 스피커', price: '25000원', stock: 10, discount: 25 },
+            { id: 'p5', name: '코딩할 때 듣는 Lo-Fi 스피커', price: '25000원', stock: 10, discount: 25 }
           ];
 
           expect(sel.options.length).toBe(5);
@@ -151,6 +150,7 @@ describe('basic 테스트', () => {
         describe('3.3.1 화요일 할인', () => {
           it('화요일에 10% 추가 할인 적용', () => {
             const tuesday = new Date('2024-10-15'); // 화요일
+
             vi.setSystemTime(tuesday);
 
             sel.value = 'p1';
@@ -163,10 +163,13 @@ describe('basic 테스트', () => {
             // 화요일 특별 할인 배너 표시
             const tuesdayBanner = document.getElementById('tuesday-special');
             expect(tuesdayBanner.classList.contains('hidden')).toBe(false);
+
+
           });
 
           it('화요일 할인은 다른 할인과 중복 적용', () => {
             const tuesday = new Date('2024-10-15');
+
             vi.setSystemTime(tuesday);
 
             addItemsToCart(sel, addBtn, 'p1', 10);
@@ -174,44 +177,56 @@ describe('basic 테스트', () => {
             // 100,000원 -> 90,000원 (개별 10%) -> 81,000원 (화요일 10% 추가)
             expect(sum.textContent).toContain('₩81,000');
             expect(discountInfo.textContent).toContain('19.0%'); // 총 19% 할인
+
+
           });
         });
 
         describe('3.3.2 번개세일', () => {
           it.skip('번개세일 알림 표시 및 20% 할인 적용', async () => {
             // 원본 코드의 타이머 구현 문제로 인해 스킵
+
             await vi.advanceTimersByTimeAsync(40000);
+
           });
 
           it.skip('번개세일 상품은 드롭다운에 ⚡ 아이콘 표시', async () => {
             // 원본 코드의 타이머 구현 문제로 인해 스킵
+
             await vi.advanceTimersByTimeAsync(40000);
+
           });
         });
 
         describe('3.3.3 추천할인', () => {
           it.skip('마지막 선택한 상품과 다른 상품 추천 및 5% 할인', async () => {
             // 원본 코드의 타이머 구현 문제로 인해 스킵
+
             sel.value = 'p1';
             addBtn.click();
             await vi.advanceTimersByTimeAsync(80000);
+
           });
 
           it.skip('추천할인 상품은 드롭다운에 💝 아이콘 표시', async () => {
             // 원본 코드의 타이머 구현 문제로 인해 스킵
+
             sel.value = 'p1';
             addBtn.click();
             await vi.advanceTimersByTimeAsync(80000);
+
           });
         });
 
         describe('3.3.4 할인 중복', () => {
           it.skip('번개세일 + 추천할인 = 25% SUPER SALE', async () => {
             // 원본 코드의 타이머 구현 문제로 인해 스킵
+
             await vi.advanceTimersByTimeAsync(40000);
             sel.value = 'p1';
             addBtn.click();
             await vi.advanceTimersByTimeAsync(80000);
+
           });
         });
       });
@@ -232,6 +247,7 @@ describe('basic 테스트', () => {
       describe('4.2 추가 적립', () => {
         it('화요일 구매 시 기본 포인트 2배', () => {
           const tuesday = new Date('2024-10-15');
+
           vi.setSystemTime(tuesday);
 
           sel.value = 'p1';
@@ -240,6 +256,8 @@ describe('basic 테스트', () => {
           // 9,000원 (화요일 10% 할인) -> 9포인트 * 2 = 18포인트
           expect(loyaltyPoints.textContent).toContain('18p');
           expect(loyaltyPoints.textContent).toContain('화요일 2배');
+
+
         });
 
         it('키보드+마우스 세트 구매 시 +50p', () => {
@@ -337,7 +355,7 @@ describe('basic 테스트', () => {
           const options = Array.from(sel.options);
 
           // 품절 상품이 비활성화되어 있는지 확인
-          const disabledOption = options.find((opt) => opt.disabled);
+          const disabledOption = options.find(opt => opt.disabled);
           if (disabledOption) {
             expect(disabledOption.textContent).toContain('품절');
           }
@@ -612,6 +630,7 @@ describe('basic 테스트', () => {
           sel.value = 'p1';
           addBtn.click();
           await vi.advanceTimersByTimeAsync(80000);
+
         });
       });
     });
@@ -633,17 +652,22 @@ describe('basic 테스트', () => {
 
         // 포인트 확인: 405포인트(기본) * 2(화요일) + 50(세트) + 100(풀세트) + 100(30개) = 1060포인트
         expect(loyaltyPoints.textContent).toContain('1060p');
+
+
       });
 
       it.skip('번개세일 + 추천할인 + 화요일 시나리오', async () => {
         // 원본 코드의 타이머 구현 문제로 인해 스킵
         const tuesday = new Date('2024-10-15');
+
         vi.setSystemTime(tuesday);
 
         await vi.advanceTimersByTimeAsync(40000);
         sel.value = 'p1';
         addBtn.click();
         await vi.advanceTimersByTimeAsync(80000);
+
+
       });
     });
   });
